requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
python-dotenv
psycopg2-binary
streamlit
requests
langchain_community